requirements_test.txt
mock==1.0.1
vcrpy==1.4.1
nose
coverage
unittest2